+++ bitbots_localization_handler/bitbots_localization_handler/localization_dsd/decisions/game_state.py
@@ -1,105 +1,6 @@
-import numpy as np
-import tf2_ros as tf2
-from bitbots_msgs.msg import GameState, RobotControlState
-from rclpy.time import Time
+from bitbots_msgs.msg import GameState
 
 from bitbots_localization_handler.localization_dsd.decisions import AbstractLocalizationDecisionElement
-
-
-class CheckFallen(AbstractLocalizationDecisionElement):
-    """
-    Checks if robot is fallen
-    """
-
-    def perform(self, reevaluate=False):
-        self.clear_debug_data()
-
-        if self.blackboard.robot_control_state == RobotControlState.FALLEN:
-            return "FALLEN"
-        return "NOT_FALLEN"
-
-    def get_reevaluate(self):
-        return True
-
-
-class CheckFalling(AbstractLocalizationDecisionElement):
-    """
-    Checks if robot is falling
-    """
-
-    def perform(self, reevaluate=False):
-        self.clear_debug_data()
-
-        if self.blackboard.robot_control_state == RobotControlState.FALLING:
-            return "FALLING"
-        return "NOT_FALLING"
-
-    def get_reevaluate(self):
-        return True
-
-
-class CheckGettingUp(AbstractLocalizationDecisionElement):
-    """
-    Checks if robot is getting up
-    """
-
-    def perform(self, reevaluate=False):
-        self.clear_debug_data()
-
-        if self.blackboard.robot_control_state == RobotControlState.GETTING_UP:
-            return "GETTING_UP"
-        return "NOT_GETTING_UP"
-
-    def get_reevaluate(self):
-        return True
-
-
-class CheckPickup(AbstractLocalizationDecisionElement):
-    """
-    Checks if robot is picked up
-    """
-
-    def perform(self, reevaluate=False):
-        self.clear_debug_data()
-
-        if self.blackboard.robot_control_state == RobotControlState.PICKED_UP:
-            self.blackboard.last_state_pickup = True
-            return "UP"
-        else:
-            if self.blackboard.last_state_pickup:
-                self.blackboard.last_state_pickup = False
-                return "JUST_DOWN"
-
-        return "DOWN"
-
-    def get_reevaluate(self):
-        return True
-
-
-class GettingUpState(AbstractLocalizationDecisionElement):
-    """
-    Checks if the robot falls, stands up or is freshly standing
-    """
-
-    def __init__(self, blackboard, dsd, parameters=None):
-        super().__init__(blackboard, dsd, parameters)
-        self.get_up_states = [RobotControlState.FALLING, RobotControlState.FALLEN, RobotControlState.GETTING_UP]
-
-    def perform(self, reevaluate=False):
-        self.clear_debug_data()
-
-        if self.blackboard.robot_control_state in self.get_up_states:
-            self.blackboard.last_state_get_up = True
-            return "YES"
-        else:
-            if self.blackboard.last_state_get_up:
-                self.blackboard.last_state_get_up = False
-                return "GOTUP"
-
-        return "NO"
-
-    def get_reevaluate(self):
-        return True
 
 
 class CheckGameStateReceived(AbstractLocalizationDecisionElement):
@@ -125,16 +26,6 @@ class CheckGameStateReceived(AbstractLocalizationDecisionElement):
 
 
 class GameStateDecider(AbstractLocalizationDecisionElement):
-    def __init__(self, blackboard, dsd, parameters=None):
-        super().__init__(blackboard, dsd, parameters)
-        self.game_states = {
-            0: "INITIAL",
-            1: "READY",
-            2: "SET",
-            3: "PLAYING",
-            4: "FINISHED",
-        }
-
     def perform(self, reevaluate=False):
         """
         Translates GameState in Blackboard into DSD Answer
@@ -167,23 +58,9 @@ class SecondaryStateDecider(AbstractLocalizationDecisionElement):
     game controller receiver, so the behavior does ont need to deal with this.
     """
 
-    def __init__(self, blackboard, dsd, parameters=None):
-        super().__init__(blackboard, dsd, parameters)
-        self.secondary_game_states = {
-            0: "NORMAL",
-            1: "PENALTYSHOOT",  # should not happen during halftime or extra time
-            2: "OVERTIME",
-            3: "TIMEOUT",
-            4: "DIRECT_FREEKICK",
-            5: "INDIRECT_FREEKICK",
-            6: "PENALTYKICK",
-            7: "CORNER_KICK",
-            8: "GOAL_KICK",
-            9: "THROW_IN",
-        }
-
     def perform(self, reevaluate=False):
         state_number = self.blackboard.gamestate.get_secondary_state()
+
         # todo this is a temporary hack to make GUI work
         if state_number == GameState.STATE_NORMAL:
             return "NORMAL"
@@ -242,9 +119,6 @@ class SecondaryStateTeamDecider(AbstractLocalizationDecisionElement):
 
 
 class CheckPenalized(AbstractLocalizationDecisionElement):
-    def __init__(self, blackboard, dsd, parameters=None):
-        super().__init__(blackboard, dsd, parameters)
-
     def perform(self, reevaluate=False):
         """
         Determines if the robot is penalized by the game controller.
@@ -259,56 +133,6 @@ class CheckPenalized(AbstractLocalizationDecisionElement):
             return "NO"
 
     def get_reevaluate(self):
-        return True
-
-
-class WalkedSinceLastInit(AbstractLocalizationDecisionElement):
-    """
-    Decides if we walked significantly since our last initialization
-    """
-
-    def __init__(self, blackboard, dsd, parameters=None):
-        super().__init__(blackboard, dsd, parameters)
-        self.distance_threshold = parameters.get("dist", 0.5)
-
-    def perform(self, reevaluate=False):
-        if not self.blackboard.use_sim_time:
-            # in real life we always have moved and are not teleported
-            return "YES"
-
-        if self.blackboard.last_init_odom_transform is None:
-            return "YES"  # We don't know the last init state so we say that we moved away from it
-
-        try:
-            odom_transform = self.blackboard.tf_buffer.lookup_transform(
-                self.blackboard.odom_frame, self.blackboard.base_footprint_frame, Time(seconds=0, nanoseconds=0)
-            )
-        except (tf2.LookupException, tf2.ConnectivityException, tf2.ExtrapolationException) as e:
-            self.blackboard.node.get_logger().error(
-                f"Reset localization to last init state, because we got up and have no tf: {e}"
-            )
-            # We assume that we didn't walk if the tf lookup fails
-            return "YES"
-
-        walked_distance = np.linalg.norm(
-            np.array([odom_transform.transform.translation.x, odom_transform.transform.translation.y])
-            - np.array(
-                [
-                    self.blackboard.last_init_odom_transform.transform.translation.x,
-                    self.blackboard.last_init_odom_transform.transform.translation.y,
-                ]
-            )
-        )
-
-        if walked_distance < self.distance_threshold:
-            return "NO"
-        else:
-            return "YES"
-
-    def get_reevaluate(self):
-        """
-        The state can change during the game
-        """
         return True
